+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## [1.4.2](https://github.com/adobe/helix-fetch/compare/v1.4.1...v1.4.2) (2020-03-24)
+
+
+### Bug Fixes
+
+* **index:** sanitize method name ([fc539e1](https://github.com/adobe/helix-fetch/commit/fc539e1cf740a148e0371b76dbbf69811bdf1d5f)), closes [#24](https://github.com/adobe/helix-fetch/issues/24)
+
+## [1.4.1](https://github.com/adobe/helix-fetch/compare/v1.4.0...v1.4.1) (2020-03-17)
+
+
+### Performance Improvements
+
+* bumped fetch-h2 dependency and some housekeeping ([db66e1e](https://github.com/adobe/helix-fetch/commit/db66e1e0240ac7626ff911ea970e073d9f4e3a45)), closes [#14](https://github.com/adobe/helix-fetch/issues/14)
+
+# [1.4.0](https://github.com/adobe/helix-fetch/compare/v1.3.0...v1.4.0) (2020-02-28)
+
+
+### Features
+
+* make cache size limit configurable ([7fd632d](https://github.com/adobe/helix-fetch/commit/7fd632df0f875b701156132be58e68b8f9c37526)), closes [#9](https://github.com/adobe/helix-fetch/issues/9)
+
+# [1.3.0](https://github.com/adobe/helix-fetch/compare/v1.2.0...v1.3.0) (2020-02-28)
+
+
+### Features
+
+* option to force HTTP/1(.1) protocol ([501d6a2](https://github.com/adobe/helix-fetch/commit/501d6a2fb08fe08657238ea4ee7e558caef392fd)), closes [#8](https://github.com/adobe/helix-fetch/issues/8)
+
+# [1.2.0](https://github.com/adobe/helix-fetch/compare/v1.1.0...v1.2.0) (2020-02-27)
+
+
+### Features
+
+* expose fetch context to allow custom configuration ([6be0b6c](https://github.com/adobe/helix-fetch/commit/6be0b6cdb172702d2f4068cc632dd9f2643072da)), closes [#15](https://github.com/adobe/helix-fetch/issues/15)
+
 # [1.1.0](https://github.com/adobe/helix-fetch/compare/v1.0.0...v1.1.0) (2020-02-03)